Fixed typo ViewFactory and changed addTaskList function for testing

+++ client/src/main/java/client/utils/ServerUtils.java
@@ -19,7 +19,6 @@ import jakarta.ws.rs.client.ClientBuilder;
 import jakarta.ws.rs.client.Entity;
 import jakarta.ws.rs.core.GenericType;
 import jakarta.ws.rs.core.Response;
-import javafx.concurrent.Task;
 import models.Board;
 import models.TaskCard;
 import models.TaskList;
@@ -111,7 +110,7 @@ public class ServerUtils {
                 .post(Entity.entity(board,APPLICATION_JSON), Board.class);
     }
 
-    public TaskList addTaskList(TaskList taskList, int boardId) {
+    public TaskList addTaskList(TaskList taskList, Long boardId) {
         return ClientBuilder.newClient(new ClientConfig())
                 .target(SERVER).path("api/boards/taskLists/" + boardId)
                 .request(APPLICATION_JSON)
